requirements.txt
requests
beautifulsoup4
pandas
pillow
sklearn
numpy
networkx
matplotlib
skimage
